MS-74 move Employee into separate package
