+++ src/main/java/de/szut/lf8_project/domain/adapter/EmployeeRepository.java
@@ -1,8 +1,8 @@
 package de.szut.lf8_project.domain.adapter;
 
 import de.szut.lf8_project.common.JWT;
-import de.szut.lf8_project.domain.Employee;
-import de.szut.lf8_project.domain.EmployeeId;
+import de.szut.lf8_project.domain.Employee.Employee;
+import de.szut.lf8_project.domain.Employee.EmployeeId;
 import de.szut.lf8_project.repository.RepositoryException;
 
 public interface EmployeeRepository {
